# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Newsight API
  description: REST API documentation for the Newsight backend.
  contact:
    name: Newsight
  version: v1
servers:
- url: http://localhost:8080
  description: Local backend
- url: https://newsightkr.com
  description: Production
...
paths:
...
  /api/analytics/keywords/{keyword_seq}/wordcloud/title/:
    get:
      tags:
      - Keyword Analytics
      summary: Get title wordcloud
      operationId: getTitleWordcloud
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WordcloudResponseDto"
  /api/analytics/keywords/{keyword_seq}/wordcloud/title:
    get:
      tags:
      - Keyword Analytics
      summary: Get title wordcloud
      operationId: getTitleWordcloud_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WordcloudResponseDto"
  /api/analytics/keywords/{keyword_seq}/wordcloud/comment:
    get:
      tags:
      - Keyword Analytics
      summary: Get comment wordcloud
      operationId: getCommentWordcloud
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WordcloudResponseDto"
  /api/analytics/keywords/{keyword_seq}/wordcloud/comment/:
    get:
      tags:
      - Keyword Analytics
      summary: Get comment wordcloud
      operationId: getCommentWordcloud_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WordcloudResponseDto"
  /api/analytics/keywords/{keyword_seq}/summary:
    get:
      tags:
      - Keyword Analytics
      summary: Get AI summary
      operationId: getAiSummary
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AiSummaryResponseDto"
  /api/analytics/keywords/{keyword_seq}/summary/:
    get:
      tags:
      - Keyword Analytics
      summary: Get AI summary
      operationId: getAiSummary_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AiSummaryResponseDto"
  /api/analytics/keywords/{keyword_seq}/sentiment/content:
    get:
      tags:
      - Keyword Analytics
      summary: Get content sentiment
      operationId: getContentSentiment
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContentSentimentResponseDto"
  /api/analytics/keywords/{keyword_seq}/sentiment/content/:
    get:
      tags:
      - Keyword Analytics
      summary: Get content sentiment
      operationId: getContentSentiment_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ContentSentimentResponseDto"
  /api/analytics/keywords/{keyword_seq}/search-timeline:
    get:
      tags:
      - Keyword Analytics
      summary: Get search interest timeline
      operationId: getSearchTimeline
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SearchTimelineResponseDto"
  /api/analytics/keywords/{keyword_seq}/search-timeline/:
    get:
      tags:
      - Keyword Analytics
      summary: Get search interest timeline
      operationId: getSearchTimeline_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SearchTimelineResponseDto"
  /api/analytics/keywords/{keyword_seq}/cooc-network/:
    get:
      tags:
      - Keyword Analytics
      summary: Get co-occurrence network
      operationId: getCoocNetwork
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CoocNetworkResponseDto"
  /api/analytics/keywords/{keyword_seq}/cooc-network:
    get:
      tags:
      - Keyword Analytics
      summary: Get co-occurrence network
      operationId: getCoocNetwork_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CoocNetworkResponseDto"
  /api/analytics/keywords/{keyword_seq}/bias/title:
    get:
      tags:
      - Keyword Analytics
      summary: Get title bias by media
      operationId: getTitleBiasByMedia
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TitleBiasByMediaResponseDto"
  /api/analytics/keywords/{keyword_seq}/bias/title/:
    get:
      tags:
      - Keyword Analytics
      summary: Get title bias by media
      operationId: getTitleBiasByMedia_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TitleBiasByMediaResponseDto"
  /api/analytics/keywords/{keyword_seq}/:
    get:
      tags:
      - Keyword Analytics
      summary: Get keyword metadata
      operationId: getKeywordMeta
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KeywordMetaResponseDto"
  /api/analytics/keywords/{keyword_seq}:
    get:
      tags:
      - Keyword Analytics
      summary: Get keyword metadata
      operationId: getKeywordMeta_1
      parameters:
      - name: keyword_seq
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: period
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KeywordMetaResponseDto"
...
components:
  schemas:
...
    WordItemDto:
      type: object
      properties:
        word:
          type: string
        weight:
          type: number
          format: double
    WordcloudResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WordItemDto"
    AiSummaryResponseDto:
      type: object
      properties:
        summary_text:
          type: string
    ContentSentimentResponseDto:
      type: object
      properties:
        positive:
          type: number
          format: double
        neutral:
          type: number
          format: double
        negative:
          type: number
          format: double
    SearchTimelineResponseDto:
      type: object
      properties:
        period_start:
          type: string
        period_end:
          type: string
        latest_score:
          type: integer
          format: int32
        peak_score:
          type: integer
          format: int32
        average_score:
          type: number
          format: double
        has_partial:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/TimelinePointDto"
    TimelinePointDto:
      type: object
      properties:
        observed_date:
          type: string
        interest_score:
          type: integer
          format: int32
        is_partial:
          type: boolean
    CoocNetworkResponseDto:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeItemDto"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/EdgeItemDto"
    EdgeItemDto:
      type: object
      properties:
        source:
          type: integer
          format: int64
        target:
          type: integer
          format: int64
        weight:
          type: number
          format: double
    NodeItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        size:
          type: number
          format: double
    BiasByMediaItemDto:
      type: object
      properties:
        media_name:
          type: string
        bias_score:
          type: number
          format: double
    TitleBiasByMediaResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BiasByMediaItemDto"
    KeywordMetaResponseDto:
      type: object
      properties:
        keyword_seq:
          type: integer
          format: int64
        keyword:
          type: string
        period_start:
          type: string
        period_end:
          type: string
        article_count:
          type: integer
          format: int32
        media_count:
          type: integer
          format: int32
        is_analyzable:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Use the access token returned by the login API.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT